Moved utilities to a submodule.
Darrays were moved into the utilities repo. Vectors and procedural generation are no loonger in the sgl namespace but are in the util namespace.
The utilities include and using directive for the util namespace was added to the sgl include file so that the changes to the codebase were minimal.