Removed expression trees.
Expression trees are no longer in the SGL namespace but are implemented in the util namespace as part of the Utility submodule.
Expression trees are no longer a part of the render tree model. Render functions are now expected to return a single RGBAD value immediately. Expression trees are no longer referenced anywhere in Voxel's code.